Makefile: add an option to rename dirdiff
The EXE variable in Makefile can be used to change the name of the
executable referenced by the man page and command completion scripts.
The executable itself retains its name which can be changed trivially
with `mv`.

This change aims to help avoid a name collision of packages on e.g.
Debian, where a package named `dirdiff` is already in the repo. It will
also make it easier to install multiple versions of dirdiff at the same
time on one system e.g. for testing the beta releases.