rename dirdiff to durduff
The name "dirdiff" is already taken by a package in the official Debian
repos. "durduff" is a pun on this name. When searching on GitHub, I
found no collisions with this name, so there is a good chance that with
this name we won't encounter collisions in other package repos.