add NEWS file
The NEWS file is where the changelog will be kept. Its format is inspired
by Drupal's CHANGELOG.txt. The filename "NEWS" is widely-used for this
purpose in the open-source world.

The NEWS file should be plain-text, because it will often be read in a
terminal (using a pager, a text editor, APT, etc).