configure deb package auto-generation
In order to deliver new releases of durduff in a way convenient for
end-users and to integrate with the target platforms, durduff needs to
be available in packages in formats pre-established on these platforms.

On Debian-based Linux distributions this format is DEB packages. It
allows for easy installation, updates and uninstallation.

The 'deb' Makefile target generates any needed assets (at this point
these are the compressed man page and changelog) and uses 'cargo-deb' to
generate a new DEB package according to the options specified in
section 'package.metadata.deb' of Cargo.toml.

As of this commit, Lintian v2.15.0 doesn't have any complaints about the
result.