move git-specific parts of Makefile to with-git.mk
When somebody uses the source tarball to build durduff, they shouldn't
see any errors caused by the lack of ".git" directory. Some of the
variables defined in Makefile, e.g. last_tag, are filled with the output
of git commands. They would be evaluated each time a make command is
run.

"with-git.mk" will be dynamically included by Makefile if (and only if)
the ".git" directory is present.