README.md: drop TODO of Mac packages
README.md used to declare that Mac packages are a TODO. However, recent
developments[1][2] in Apple-land prompted me to reconsider my willingness
to support Mac in any capacity at all.

I am not going to pay $100 a year just to enable users of an oppresive
system to run the software I write. Neither do I think that others
should. Instead this system deserves to die[3].

Apple's dropping of OpenGL support and push for proprietary APIs point
me in the same direction. Although it's irrelevant for this little
project, it is relevant for other projects I have planned for the near
and more distant future; this project will share some code and
infrastructure around CI and release process. Thus its future is closely
tied with projects that may depend on OpenGL.

For the past several weeks, on and off, I've been reading about the
possibilities of setting up a CI pipeline for building and testing
software for Mac. My impression is that it would be nowhere near as easy
as it is with Linux or BSDs. The main difficulty stems of course from the
fact that in some jurisdictions (US) running macOS in a VM on non-Mac
hardware is illegal, and in some jurisdictions (EU) the legal status of
doing it is unclear.

[1]: <https://developer.apple.com/news/?id=09032019a>
[2]: <http://www.nathalielawhead.com/candybox/the-future-of-my-games-on-apple-post-catalina-and-what-this-means-for-art-games-in-general>
[3]: I am under no illusions that the availability of my software
     has any significance in the matter.