Make percent-encoding the default output format
It makes the output more deterministic --- there is no need to check if
the preceding line shows an error about not being able to print a given
file name in a single line.

For scripting purposes I added NUL-terminated output (enabled with the
--null option). This way scripts will not need to implement
percent-decoding, it should be faster and it already is a common unix
CLI idiom (see find(1), xargs(1), grep(1)).

Windows support is abandoned. The main reason for that is that on
Windows filenames are case-insensitive and I couldn't find any docs on
what Windows considers to be the correct uppercase->lowercase mapping.
Without applying any kind of case-normalization, durduff would be buggy
on Windows. It seems that there is a function that can be called from
C#, but native languages are out of luck. On top of that, Microsoft
added to Windows a thing called "Per-directory case sensitivity"[1] and
at this point any sanity that was left is thrown out the window. All in
all, Windows isn't a very friendly platform to develop for when you're
dealing with files *facepalm*.

[1]: <https://devblogs.microsoft.com/commandline/per-directory-case-sensitivity-and-wsl/>

At this point, NUL-terminated output isn't covered in the functional
tests. This is because entering NUL-terminated text into the input
(func.test file) of shelltestrunner isn't all that easy and future
modifications would be even more tricky. I could probably rewrite the
tests as classic Rust tests, but the result would look pretty ugly and
I'm lazy. I may end up filing a bug report to shelltestrunner about
that, so they implement escaped input (so I can just write '\0' for the
NUL character) and maybe I'll even implement that myself. We'll see. For
now this output format isn't tested. Let's hope everything works.