Rewrite functional tests into Rust #tests
Now functional tests don't require shelltestrunner. Moreover, tests for
the NUL line terminator were added (which wasn't possible with
shelltestrunner).

There are some other minor changes included, because along the way I did
a lot of exploring of what I could do with this program and how well it
would work. Most of that didn't make it into the rewrite, because I
prefer to do less, but reliably, than do more, but break things from
time to time.

One of those things was e.g. APT-style progress bar. Implementing it
requires handling Unix signals which would complicate code a lot. I
actually got that part working, but then on top of that I discovered
that APT's progress bar doesn't play well with terminal resizing
(particularly making the terminal window narrower). And I don't think
that this type of a progress bar can be made to work, given the way that
terminals work these days (ansi escape sequences work on visual lines
and not on semantic lines; saving and restoring cursor position only
operates on a single "cell" in memory --- there is no way to store
multiple positions).

As it is now, a lot of functions, structs and traits lack documentation.
I'm too tired to do it now. I'll add it in subsequent commits.