RecordPrinter: Flush as soon as we write to stderr
It's important that when stdout is the same thing as stderr (as is the
case when durduff is used interactively from a terminal without
redirecting stdout nor stderr) and we encounter an error and print
"^ {path}" to stdout, the next line of stderr describes the error and
that nothing is printed between "^ {path}" and the description of the
encountered error. This requires that we flush stdout and stderr as soon
as anything is written to the stderr buffer.